[API-91] Extract manual/schedule/system/notification-settings/replan/zones-summary route modules.
